fix: 기존 컨트롤러를 관리자 컨트롤러로 분기.
