+++ src/main/java/io/chaerin/cafemanagement/domain/order/controller/OrderController.java
@@ -6,10 +6,6 @@ import io.chaerin.cafemanagement.domain.order.dto.OrderUpdateRequestDto;
 import io.chaerin.cafemanagement.domain.order.service.OrderService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +13,6 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @Controller
-
 @RequestMapping("/orders")
 @RequiredArgsConstructor
 public class OrderController {
@@ -57,27 +52,5 @@ public class OrderController {
         return "redirect:/orders";
     }
 
-    @GetMapping("/admin")
-    public String getAllOrders(HttpSession session,Model model, @PageableDefault(size = 10, sort = "createdAt", direction = Sort.Direction.ASC) Pageable pageable) {
-        Page<OrderResponseDto> orders = orderService.getAllOrders(pageable);
-        if(orders.isEmpty()) {
-            return "redirect:/";
-        }
-        model.addAttribute("orders", orders);
-        return "order/adminList";
-    }
 
-    @PutMapping("/admin/{id}")
-    public String updateOrderAdmin(@PathVariable Long id, @ModelAttribute OrderUpdateRequestDto request, Model model) {
-        OrderResponseDto updatedOrder = orderService.updateOrder(id, request);
-        model.addAttribute("order", updatedOrder);
-        // 임의지정
-        return "redirect:/orders/admin";
-    }
-
-    @DeleteMapping("/admin/{id}")
-    public String deleteOrderAdmin(@PathVariable Long id) {
-        orderService.deleteOrder(id);
-        return "redirect:/orders/admin";
-    }
 }
